+++ scripts/test-helper.sh
@@ -175,6 +175,7 @@ clock
 crypto-cbc
 crypto-ccm
 crypto-ec
+crypto-ecdh
 crypto-ecdsa
 crypto-gcm
 crypto-hash
@@ -204,8 +205,10 @@ crypto-aes256-gcm
 crypto-hmac-sha256
 crypto-hmac-sha384
 crypto-p256
+crypto-p256-ecdh
 crypto-p256-ecdsa
 crypto-p384
+crypto-p384-ecdh
 crypto-p384-ecdsa
 crypto-sha256
 crypto-sha384
@@ -227,8 +230,10 @@ aes256-gcm
 hmac-sha256
 hmac-sha384
 p256
+p256-ecdh
 p256-ecdsa
 p384
+p384-ecdh
 p384-ecdsa
 sha256
 sha384